Addition of BaseFFXTest and PJDependentTest for generic FFX tests (setting logging appropriately) and FFX tests requiring the PJ world communicator, respectively. Subclass these, and beforeClass and afterClass will be called automatically. Also, Travis temporarily reconfigured for test classes to print out fine-logging messages.
