+++ modules/algorithms/pom.xml
@@ -108,7 +108,6 @@
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
             <version>${junit.version}</version>
-            <scope>test</scope>
             <type>jar</type>
         </dependency>
     </dependencies>
