+++ modules/potential/src/test/java/ffx/potential/groovy/EnergyTest.java
@@ -39,12 +39,9 @@ package ffx.potential.groovy;
 
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 
-import org.junit.AfterClass;
+import ffx.utilities.BaseFFXTest;
 import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
@@ -62,7 +59,7 @@ import groovy.lang.Binding;
  * Test the Energy script.
  */
 @RunWith(Parameterized.class)
-public class EnergyTest {
+public class EnergyTest extends BaseFFXTest {
 
     @Parameters
     public static Collection<Object[]> data() {
@@ -333,29 +330,6 @@ public class EnergyTest {
     private boolean ffxCI;
     private boolean ffxOpenMM;
 
-    private static final Logger logger = Logger.getLogger(EnergyTest.class.getName());
-    private static final Level origLevel = Logger.getLogger("ffx").getLevel();
-    private static final Level testLevel;
-    private static final Level ffxLevel;
-    static {
-        Level lev;
-        try {
-            lev = Level.parse(System.getProperty("ffx.test.log", "INFO").toUpperCase());
-        } catch (Exception ex) {
-            logger.warning(String.format(" Exception %s in parsing value of ffx.test.log", ex));
-            lev = origLevel;
-        }
-        testLevel = lev;
-
-        try {
-            lev = Level.parse(System.getProperty("ffx.log", "INFO").toUpperCase());
-        } catch (Exception ex) {
-            logger.warning(String.format(" Exception %s in parsing value of ffx.log", ex));
-            lev = origLevel;
-        }
-        ffxLevel = lev;
-    }
-
     public EnergyTest(String info, String filename, int nAtoms,
                       double bondEnergy, int nBonds,
                       double angleEnergy, int nAngles,
@@ -415,19 +389,6 @@ public class EnergyTest {
         binding = new Binding();
 
         System.clearProperty("platform");
-    }
-
-    @BeforeClass
-    public static void beforeClass() {
-        // Set appropriate logging levels for interior/exterior Loggers.
-        Logger.getLogger("ffx").setLevel(ffxLevel);
-        logger.setLevel(testLevel);
-    }
-
-    @AfterClass
-    public static void afterClass() {
-        Logger.getLogger("ffx").setLevel(origLevel);
-        logger.setLevel(origLevel);
     }
 
     @Test
