+++ modules/utilities/pom.xml
@@ -44,7 +44,6 @@
             <artifactId>junit</artifactId>
             <version>${junit.version}</version>
             <type>jar</type>
-            <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>commons-math</groupId>
